relocate WavesurferWaveform
