+++ src/components/Wavesurfer/WavesurferWaveform.tsx
